+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * **New** - Elementor Dynamic Tag: Book Package Price
 * **Tweak** - Make ACF field pitch translatable
 * **Tweak** - Make ACF field secondary_desc translatable
+* **New** - ACF new relation for Trip Type: Accommodation Type
 
 ### 1.1.1: August 30th, 2018
 * **Tweak** - Change the way ACF Fields are loaded (HPP replaced JSON)